requirements.txt
asgiref               3.5.2
backports.zoneinfo    0.2.1
Django                4.1
django-cors-headers   3.13.0
django-simple-history 3.1.1
djangorestframework   3.13.1
MarkupSafe            2.1.1
mysqlclient           2.1.1
pip                   20.2.3
PyJWT                 2.4.0
pytz                  2022.2.1
setuptools            49.2.1
sqlparse              0.4.2
tzdata                2022.2
Werkzeug              2.2.2